+++ include/Sprite.hpp
@@ -120,7 +120,7 @@ public:
 	 * @param y
 	 *   Pixel drawing position on vertical axis.
 	 */
-	void render(ViewportRenderer* viewport, uint16_t x, uint16_t y);
+	void render(ViewportRenderer* viewport, uint32_t x, uint32_t y);
 };
 
 #endif /* RRE_SPRITE_H */
